Remove accidentaly `compileOnly extendsFrom compile`
This was an oversight when we implemented compileOnly dependencies.
In a first iteration `compileOnly` was set as the sourceSet's compile classpath.
We later changed this and introduced the `compileClasspath` configuration to better
split these concerns. At that point we forgot to remove the extends relationship
from compileOnly. The result is that there is no convenient way to query for all
dependencies that are indeed "compile only".

After this change users can now resolve `compileOnly` and actually only get the
dependencies that are used during compilation only, instead of also getting the
`compile` dependencies.